Combine event routes into one file
